kasper: Adds snapshotting between tests.
Only works with qemu on supported architectures. Future work is to
support enabling/disabling snapshotting, but for now it's always
enabled.

The snapshotting works as follows. All communication between the fuzzer
and manager is done via RPC unless otherwise noted:
1. Before running each test, the fuzzer closes the RPC session.
2. The fuzzer sends a transient, asynchronous SaveSnapshot command to
the manager.
3. The manager saves a snapshot.
4. The fuzzer queries the manager via transient, synchronous calls
until the manager indicates that the snapshot operation is finished.
5. Once it is finished, the fuzzer initiates a new RPC connection.
6. The manager indicates to the fuzzer that it should enable kasper
for this test (as it is the first test run).
7. The fuzzer enables kasper then runs the test case.
8. The fuzzer closes its RPC session.
9. The fuzzer sends a transient, asynchronous LoadSnapshot command to
the manager.
10. The manager loads the snapshot.
11. Steps 4 through 7 are repeated except kasper is not enabled (as
it is the second test run).
12. The fuzzer moves on to the next test and back to step 1.

Commands from the manager (e.g., to run the fuzzer) now run in the
background, with stdout/stderr forwarded to the console. This is because
running them in the foreground persists the SSH connection, and when a
snapshot is loaded, the SSH session disconnects.